test: refactor imagetool tests
