+++ tests/conftest.py
@@ -1,11 +1,15 @@
 import os
 import pathlib
+import time
+from collections.abc import Callable
 
 import lmfit
 import numpy as np
 import pooch
 import pytest
 import xarray as xr
+from numpy.testing import assert_almost_equal
+from qtpy import QtCore, QtWidgets
 
 from erlab.io.exampledata import generate_data_angles, generate_gold_edge
 
@@ -64,3 +68,101 @@ def gold():
     return generate_gold_edge(
         temp=100, Eres=1e-2, nx=15, ny=150, edge_coeffs=(0.04, 1e-5, -3e-4), noise=False
     )
+
+
+def _accept_dialog(
+    dialog_trigger: Callable,
+    time_out: int = 5,
+    pre_call: Callable | None = None,
+    accept_call: Callable | None = None,
+) -> None:
+    """Accept a dialog during testing.
+
+    If there is no dialog, it waits until one is created for a maximum of 5 seconds (by
+    default). Adapted from `this issue comment on pytest-qt
+    <https://github.com/pytest-dev/pytest-qt/issues/256#issuecomment-1915675942>`_.
+
+    Parameters
+    ----------
+    dialog_trigger
+        Callable that triggers the dialog creation.
+    time_out
+        Maximum time (seconds) to wait for the dialog creation.
+    pre_call
+        Callable that takes the dialog as a single argument. If provided, it is executed
+        before calling ``.accept()`` on the dialog.
+    accept_call
+        If provided, it is called instead of ``.accept()`` on the dialog.
+    """
+    dialog = None
+    start_time = time.time()
+
+    # Helper function to catch the dialog instance and hide it
+    def dialog_creation():
+        # Wait for the dialog to be created or timeout
+        nonlocal dialog
+        while dialog is None and time.time() - start_time < time_out:
+            dialog = QtWidgets.QApplication.activeModalWidget()
+
+        # Avoid errors when dialog is not created
+        if isinstance(dialog, QtWidgets.QDialog):
+            if pre_call is not None:
+                pre_call(dialog)
+
+            if accept_call is not None:
+                accept_call(dialog)
+            elif (
+                isinstance(dialog, QtWidgets.QMessageBox)
+                and dialog.defaultButton() is not None
+            ):
+                dialog.defaultButton().click()
+            else:
+                dialog.accept()
+
+    # Create a thread to get the dialog instance and call dialog_creation trigger
+    QtCore.QTimer.singleShot(1, dialog_creation)
+    dialog_trigger()
+
+    assert isinstance(
+        dialog, QtWidgets.QDialog
+    ), f"No dialog was created after {time_out} seconds. Dialog type: {type(dialog)}"
+
+
+@pytest.fixture
+def accept_dialog():
+    return _accept_dialog
+
+
+def _move_and_compare_values(qtbot, win, expected, cursor=0, target_win=None):
+    if target_win is None:
+        target_win = win
+    assert_almost_equal(win.array_slicer.point_value(cursor), expected[0])
+
+    # Move left
+    qtbot.keyClick(
+        target_win, QtCore.Qt.Key.Key_Left, QtCore.Qt.KeyboardModifier.ShiftModifier
+    )
+    assert_almost_equal(win.array_slicer.point_value(cursor), expected[1])
+
+    # Move down
+    qtbot.keyClick(
+        target_win, QtCore.Qt.Key.Key_Down, QtCore.Qt.KeyboardModifier.ShiftModifier
+    )
+    assert_almost_equal(win.array_slicer.point_value(cursor), expected[2])
+
+    # Move right
+    qtbot.keyClick(
+        target_win, QtCore.Qt.Key.Key_Right, QtCore.Qt.KeyboardModifier.ShiftModifier
+    )
+    assert_almost_equal(win.array_slicer.point_value(cursor), expected[3])
+
+    # Move up
+    qtbot.keyClick(
+        target_win, QtCore.Qt.Key.Key_Up, QtCore.Qt.KeyboardModifier.ShiftModifier
+    )
+    assert_almost_equal(win.array_slicer.point_value(cursor), expected[0])
+
+
+@pytest.fixture
+def move_and_compare_values():
+    return _move_and_compare_values
